YP_FINAL_V2/server/requirements_server.txt
fastapi
uvicorn[standard]
sqlalchemy
lyricsgenius
requests
backoff
torch
transformers
sentence-transformers
cryptography
python-dotenv
sentencepiece
scipy
nltk
huggingface_hub[hf_xet]
faiss-cpu










